pyproject.toml
[project]
name = "ntc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.3.0",
    "pillow>=11.3.0",
    "pathlib2>=2.3.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]
